requirements.txt
streamlit
openai
PyMuPDF
PyPDF2
python-dotenv
apify-client        